requirements.txt
boto3
streamlit
faiss-cpu
python-dotenv
langchain
langchain_community
langchain_aws
Pdfplumber
tiktoken
streamlit-chat
botocore
fastapi
pydantic
uvicorn